Conflicts:
	app/Config/core.php
	lib/Cake/Console/Command/TestShell.php
	lib/Cake/Core/Object.php
	lib/Cake/Network/Http/HttpSocket.php
	lib/Cake/Routing/Router.php
	lib/Cake/Test/Case/Controller/Component/EmailComponentTest.php
	lib/Cake/Test/Case/Controller/ComponentTest.php
	lib/Cake/Test/Case/Controller/ControllerMergeVarsTest.php
	lib/Cake/Test/Case/Core/ObjectTest.php
	lib/Cake/Test/Case/Model/CakeSchemaTest.php
	lib/Cake/Test/Case/Model/Datasource/Database/PostgresTest.php
	lib/Cake/Test/Case/Model/Datasource/Database/SqliteTest.php
	lib/Cake/Test/Case/Model/Datasource/Database/SqlserverTest.php
	lib/Cake/Test/Case/Model/Datasource/DboSourceTest.php
	lib/Cake/Test/Case/Network/Http/HttpSocketTest.php
	lib/Cake/Test/Case/Routing/DispatcherTest.php
	lib/Cake/Test/Case/View/ScaffoldViewTest.php
	lib/Cake/Test/TestCase/Cache/Engine/FileEngineTest.php
	lib/Cake/Test/TestCase/Cache/Engine/RedisEngineTest.php
	lib/Cake/Test/TestCase/Console/Command/BakeShellTest.php
	lib/Cake/Test/TestCase/Console/Command/Task/ControllerTaskTest.php
	lib/Cake/Test/TestCase/Console/Command/Task/ViewTaskTest.php
	lib/Cake/Test/TestCase/Controller/Component/AuthComponentTest.php
	lib/Cake/Test/TestCase/Controller/Component/PaginatorComponentTest.php
	lib/Cake/Test/TestCase/Controller/ControllerTest.php
	lib/Cake/Test/TestCase/Network/RequestTest.php
	lib/Cake/Test/TestCase/Routing/RouterTest.php
	lib/Cake/Test/TestCase/Utility/TimeTest.php
	lib/Cake/Test/TestCase/View/Helper/HtmlHelperTest.php
	lib/Cake/Utility/Time.php
	lib/Cake/bootstrap.php

+++ lib/Cake/Cache/Engine/FileEngine.php
@@ -369,7 +369,8 @@ class FileEngine extends CacheEngine {
 		$contents = new \RecursiveIteratorIterator($directoryIterator, \RecursiveIteratorIterator::CHILD_FIRST);
 		foreach ($contents as $object) {
 			$containsGroup = strpos($object->getPathName(), DS . $group . DS) !== false;
-			if ($object->isFile() && $containsGroup) {
+			$hasPrefix = strpos($object->getBaseName(), $this->settings['prefix']) === 0;
+			if ($object->isFile() && $containsGroup && $hasPrefix) {
 				unlink($object->getPathName());
 			}
 		}
